# openapi.yaml
openapi: 3.0.3
info:
  title: HILLink API
  version: 1.0.0
  description: |
    Production-ready OpenAPI starter generated from current Next.js route handlers.
    Authentication is session/cookie-based (Supabase). Admin routes require admin role.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: http://localhost:3001
    description: Local test
security:
  - cookieAuth: []
...
paths:
...
  /api/business/cancel-campaign:
    post:
      tags: [Business]
      summary: Cancel campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelCampaignRequest'
      responses:
        '200':
          description: Campaign cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/stripe/trigger-payout:
    post:
      tags: [Stripe]
      summary: Trigger payout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerPayoutRequest'
      responses:
        '200':
          description: Payout triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/stripe/webhook:
    post:
      tags: [Stripe]
      summary: Stripe webhook receiver
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/location/geocode:
    post:
      tags: [Utility]
      summary: Geocode a city and state to lat/lon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeocodeRequest'
      responses:
        '200':
          description: Geocode result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeocodeResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
  schemas:
    Uuid:
      type: string
      format: uuid
      example: 7658ce51-2694-425b-8cfb-8f5cc93cc6cb
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
...
    CancelCampaignRequest:
      type: object
      required: [campaignId]
      properties:
        campaignId:
          $ref: '#/components/schemas/Uuid'
...
    TriggerPayoutRequest:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          $ref: '#/components/schemas/Uuid'
    GeocodeRequest:
      type: object
      required: [city, state]
      properties:
        city:
          type: string
          maxLength: 120
        state:
          type: string
          maxLength: 120
    GeocodeResponse:
      type: object
      properties:
        lat:
          type: number
          nullable: true
        lon:
          type: number
          nullable: true
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'